Finish application by adding authorization to CRUD operations.
A lot of changes was made after the last commit. User accounts was added
to give users the ability to each add their own items to the database.
Admin privileges were added to enable a single account to add, edit and
delete any item and category. Code still has to be properly commented and
changed to adhere to PEP8 guidelines.